scripts/pyproject.toml
[project]
name = "scripts"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "arxiv>=2.2.0",
    "huggingface-hub>=0.34.3",
    "llama-cpp-python>=0.3.14",
    "pip>=25.1.1",
    "requests>=2.32.4",
    "scholarly>=1.7.11",
]

[dependency-groups]
dev = [
    "ruff>=0.12.0",
]
